LayoutTests/media/modern-media-controls/scrubber-support/ipad/scrubber-support-drag.html

<script src="../../../../resources/js-test-pre.js"></script>
<script src="../../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<style type="text/css" media="screen">
    
    video, #host {
        position: absolute;
        top: 0;
        left: 0;
    }

    video {
        width: 720px;
        height: 240px;
    }
    
</style>
<video src="../../../content/test.mp4" autoplay controls></video>
<div id="host"></div>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>ScrubberSupport</code> behavior when dragging the scrubber.");

const container = document.querySelector("div#host");
const media = document.querySelector("video");
const mediaController = createControls(container, media, null);

let numberOfEvents = 0;
scheduler.frameDidFire = function() {
    if (media.paused)
        return;

    media.pause();

    const input = mediaController.controls.timeControl.scrubber.children[1].element;
    const bounds = input.getBoundingClientRect();

    if (bounds.width === 0)
        return;

    media.addEventListener("timeupdate", () => {
        numberOfEvents++;

        debug("timeupdate event with media.currentTime = " + media.currentTime);
    });

    const minX = bounds.left + 7;
    const dragStartX = minX;
    const dragEndX = minX + bounds.width / 2;
    const centerY = bounds.top + bounds.height / 2;

    const point = document.body.appendChild(document.createElement("div"));
    point.style.left = minX + "px";
    point.style.top = centerY + "px";
    point.style.width = "1px";
    point.style.height = "1px";
    point.style.position = "absolute";
    point.style.backgroundColor = "red";

    debug(`
        uiController.dragFromPointToPoint(${dragStartX}, ${centerY}, ${dragEndX}, ${centerY}, 0.15, function() {
            uiController.uiScriptComplete("Done");
        });`);
    testRunner.runUIScript(`
        uiController.dragFromPointToPoint(${dragStartX}, ${centerY}, ${dragEndX}, ${centerY}, 0.15, function() {
            uiController.uiScriptComplete("Done");
        });`);

    scheduler.frameDidFire = null;
};

</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
