LayoutTests/media/modern-media-controls/scrubber-support/ipad/scrubber-support-drag-expected.txt

Testing the ScrubberSupport behavior when dragging the scrubber.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL media.currentTime should be within 0.01 of 3.0136. Was 0.
PASS successfullyParsed is true

TEST COMPLETE
